% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_total_emissions_year.R
\name{add_total_emissions_year}
\alias{add_total_emissions_year}
\title{Summarize: Add up total emisisons by year}
\usage{
add_total_emissions_year(ghg_em_year)
}
\arguments{
\item{ghg_em_year}{is the data frame with all of the CO2 emissions by sector, fuel type and year}
}
\value{
value in tons of CO2
}
\description{
Summarize: Add up total emisisons by year
}
\author{
Joslyn Fritz and Angie Bouche
}
