% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_ng.R
\name{average_ng}
\alias{average_ng}
\title{Summarize: Average natural gas emissions by year}
\usage{
average_ng(ghg_data)
}
\arguments{
\item{ghg}{is the data frame with all of the CO2 emissions from natural gas}
}
\value{
value in tons of CO2
}
\description{
Summarize: Average natural gas emissions by year
}
\author{
Joslyn Fritz and Angie Bouche
}
