% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_function1.R, R/impact_function2.R
\name{average_ng_cco}
\alias{average_ng_cco}
\title{Calculate the impact of CO2 as it increases the warming of the planet.}
\usage{
average_ng_cco(ghg_data)

average_ng_cco(ghg_data)
}
\arguments{
\item{ghg_data}{is the data frame with all of the CO2 emissions from all fuel types and industries}

\item{ghg_ng}{is the data frame with all of the CO2 emissions from all fuel types and industries}
}
\value{
amount of planetary warming from CO2 emissions; for every 10 tons of CO2 emitted, the planet warms by 0.1 degrees Celsius.

value in tons of CO2
}
\description{
Calculate the impact of CO2 as it increases the warming of the planet.
}
\author{
Joslyn Fritz and Angie Bouche

Joslyn Fritz and Angie Bouche
}
