% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_total_warming.R
\name{calc_total_warming}
\alias{calc_total_warming}
\title{Calculate the impact of CO2 as it increases the warming of the planet.}
\usage{
calc_total_warming(ghg)
}
\arguments{
\item{ghg_data}{is the data frame with all of the CO2 emissions from all fuel types and industries}
}
\value{
amount of planetary warming from CO2 emissions; for every 10 tons of CO2 emitted, the planet warms by 0.1 degrees Celsius.
}
\description{
Calculate the impact of CO2 as it increases the warming of the planet.
}
\author{
Joslyn Fritz and Angie Bouche
}
