% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_warming_no_coal.R
\name{calc_warming_no_coal}
\alias{calc_warming_no_coal}
\title{Calculate the impact of CO2 as it increases the warming of the planet if we replaced coal with renewable energy sources (assume no emissions)}
\usage{
calc_warming_no_coal(ghg_data)
}
\arguments{
\item{ghg_data}{is the data frame with all of the CO2 emissions from all fuel types and industries}
}
\value{
amount of planetary warming from CO2 emissions; for every 10 tons of CO2 emitted, the planet warms by 0.1 degrees Celsius.
}
\description{
Calculate the impact of CO2 as it increases the warming of the planet if we replaced coal with renewable energy sources (assume no emissions)
}
\author{
Joslyn Fritz and Angie Bouche
}
